backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
websockets==12.0
python-dotenv==1.0.0
PyGithub==2.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6
